+++ src/app/products/loan-products/view-loan-product/view-loan-product.component.ts
@@ -25,6 +25,7 @@ export class ViewLoanProductComponent implements OnInit {
   chargesDisplayedColumns: string[] = ["name", "chargeCalculationType", "amount", "chargeTimeType"];
   paymentFundSourceDisplayedColumns: string[] = ["paymentTypeId", "fundSourceAccountId"];
   feesPenaltyIncomeDisplayedColumns: string[] = ["chargeId", "incomeAccountId"];
+  isQualificationRulesPresent: boolean = false;
 
   constructor(
     private productService: ProductsService,
@@ -36,6 +37,8 @@ export class ViewLoanProductComponent implements OnInit {
   ) {
     this.route.data.subscribe((data: { loanProduct: any }) => {
       this.loanProduct = data.loanProduct;
+      this.loanProduct.prepaidAmount = this.loanProduct.terms?.prepaidAmount;
+      this.isQualificationRulesPresent = this.loanProduct?.qualificationRules != undefined;
     });
   }